requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.33.11
botocore==1.33.11
build==1.0.3
CacheControl==0.13.1
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==3.0.5
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
crashtest==0.4.1
cryptography==41.0.7
daphne==3.0.2
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.8
django-cors-headers==4.3.1
django-eventstream==4.5.1
django-grip==3.5.0
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.26.5
dulwich==0.21.7
fastjsonschema==2.19.0
filelock==3.13.1
freezegun==1.3.1
frozenlist==1.4.0
gripcontrol==4.2.0
gunicorn==21.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.0
incremental==22.10.0
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
keyring==24.3.0
kombu==5.3.4
magic-filter==1.0.12
MarkupSafe==2.1.3
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
pexpect==4.9.0
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pubcontrol==3.5.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
redis==5.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rpds-py==0.13.2
s3transfer==0.8.2
SecretStorage==3.3.3
service-identity==23.1.0
shellingham==1.5.4
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tomlkit==0.12.3
trove-classifiers==2023.11.29
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.24.0.post1
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.12
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
zope.interface==6.1
